fetch products from api

+++ InventoryManagement/ViewModels/AddProductViewModel.swift
@@ -12,14 +12,14 @@ class AddProductViewModel: ObservableObject {
     @Published var name: String = ""
     @Published var description: String = ""
     @Published var productType: ProductType = .product
-    @Published var sellingPrice: Float = 0.0
-    @Published var tax: Float = 0.0
+    @Published var sellingPrice: Double = 0.0
+    @Published var tax: Double = 0.0
     @Published var imageURL: String = ""
     @Published var category: String = ""
     @Published var items: [IMProduct] = []
     @Published var selectedPhotoData = Data()
     
-    var finalPrice: Float {
+    var finalPrice: Double {
             let totalPrice = sellingPrice + (sellingPrice * tax / 100)
             return totalPrice
         }
